--- IPCEfficiencycs.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Geekbrains.Lesson5.ScannerEmulator
-{
-    public interface IPCEfficiencycs
-    {
-        int GetCPULOad();
-        int GetMemoryLoad();
-
-    }
-}
