+++ Scanner.cs
@@ -6,17 +6,16 @@ using System.Threading.Tasks;
 
 namespace Geekbrains.Lesson5.ScannerEmulator
 {
+    public interface IScanner
+    {
+        string[] ReadScanner();
+    }
     class Scanner : IScanner
     {
         public string[] ReadScanner()
         {
-            //byte[] result = new byte[10];
-            string[] lines = System.IO.File.ReadAllLines(@"Scanner.txt");
-            //for (int i=0; i<10; i++)
-            //{
-            //    result[i]=Convert.ToByte(lines[i]);
-            //}
 
+            string[] lines = System.IO.File.ReadAllLines(@"Scanner.txt");
             return lines;
         }
